feat(bridge-withdrawer): sync logic (#1165)
## Summary
This checks the last sequencer transaction made by the bridge account
for the corresponding rollup block height in order to initialize the
watcher's event stream from the right block

## Background
The bridge withdrawer needs to know what the last finalized processed
block was.

## Changes
- [x] submitter startup will fetch the last finalized tx and parse the
rollup block out of it
- [x] watcher gets the initial rollup block height to listen to from the
submitter's startup

## Testing
The sync is part of the startup sequence, which has existing tests
already in place. This PR mounts the additional mock guards required by
the sync process to ensure all the existing tests still work.

closes #1108

---------

Co-authored-by: noot <36753753+noot@users.noreply.github.com>
Co-authored-by: elizabeth <elizabethjbinks@gmail.com>