chore(sequencer-relayer)!: remove functionality to restrict relaying blocks to only those proposed by a given validator (#1168)
## Summary
This changes the relayer to relay all blocks regardless of which
validator proposed them.

## Background
We had functionality to allow providing a validator secret key to the
relayer and an option to only relay blocks proposed by that validator.
The functionality wasn't being used anywhere, so it has been removed.

## Changes
- Removed configuration options (see below) and supporting code.

## Testing
N/A

## Breaking Changelist
- Removed env var `ASTRIA_SEQUENCER_RELAYER_VALIDATOR_KEY_FILE`.
- Removed env var
`ASTRIA_SEQUENCER_RELAYER_RELAY_ONLY_VALIDATOR_KEY_BLOCKS`.

## Related Issues
Closes #1010.