refactor(core): parse ics20 denoms as ibc or trace prefixed variants (#1181)
## Summary
Rewrites the ics20 parsing logic to distinguish between `ibc/` and
`<path>` prefixed denoms.

## Background
While https://github.com/astriaorg/astria/pull/1162 was an improvement
of what we had before, it was still problematic in that the parsing
logic did not distinguish between denom of the form
`[port/channel...]/base` and `ibc/hash`, so that a lot of the logic in
sequencer was done ad-hoc. This patch actually distinguishes between
both forms, allowing sequencer to enforce which type of denom it stores
in its database, and which type it processes further.

## Changes
- Rewrite `astria_core::primitive::v1::Denom` as an enum with variants
`TracePrefixed` and `IbcPrefixed`

## Testing
Adapt and expands unit tests, update sequencer tests.