chore(sequencer-relayer): add timeout to gRPCs to Celestia app (#1191)
## Summary
Added a five second timeout to all gRPCs targeting the Celestia app.

## Background
Currently if the Celestia app doesn't respond to any gRPC in a timely
manner, the relayer's submission flow is blocked by this. We want to
avoid this by not waiting indefinitely for any gRPC.

## Changes
- Added timeout to the gRPC channel. This is applied to all requests
sent via that channel. We can move to more fine-grained timeouts
(different per request) in the future if required.

## Testing
- Added a unit test. I opted for this approach over black box testing
for a couple of reasons: we can leverage tokio's `time::advance` to
avoid the test running for the actual duration of the timeout, and we
have a better chance to inspect the actual error returned by the RPC to
ensure it's a timeout variant.

## Related Issues
Closes #1187.