doc(conductor): update outdated spec (#1194)
## Summary
Updates the outdated and incorrect conductor spec to its current
implementation.

## Background
The old spec was horribly out of date and contained implementation
details that are not necessary to understanding its high level
functionality. The spec in this document is a first pass and barebones.
Flow charts will be added in a followup PRs.

## Changes
- update `specs/conductor.md`

## Testing
Not applicable.