feat(sequencer)!: implement `Ics20TransferDepositMemo` format for incoming ics20 transfers to bridge accounts (#1202)
## Summary
implement `Ics20TransferDepositMemo` format and handling for incoming
ics20 transfers to bridge accounts, which requires the memo be of a
specific format instead of assuming the entire memo is the destination
chain address.

## Background
the memo being just a destination chain address is not good for UX and
could lead to error.

## Changes
- implement `Ics20TransferDepositMemo` type which contains the
destination `rollup_address`
- add stateless check that the `FungibleTokenPacketData` (which is the
ics20 transfer packet itself) is not empty or too long
-  parsing and handling of the memo for incoming ics20 transfers

## Testing
unit tests

## Breaking Changelist
- the validation of the ICS20 memo is different, so this will break
previously sent ICS20 transfers to bridge accounts.