refactor(bridge-withdrawer)!: refactor startup to a separate subtask and remove balance check from startup (#1190)
## Summary
This refactors some of the startup logic into a separate startup
subtask. It also removes the minimum expected balance check from the
startup logic, as that is not necessary.

## Background
The original startup logic had the `Submitter` running parts of the
startup routine that weren't relevant to it. this moves that logic out
to the `Startup` task.

The balance check is removed because if a sequencer transaction with
insufficient balance is submitted it will just fail execution and the
service will restart. Requiring the check in the startup logic results
in clunky operation of the service, since launching it before funding it
will cause it to crash.

## Changes
- Introduce `Startup` object and task
- `Watcher` and `Submitter` now wait for the `Startup` to update the
global state object with startup info before running their startup
logic, whereas before we just had the `Watcher` wait on the `Submitter`
to send a message on the `oneshot` channel.
- Cleaned up some of the tryhard configs with helper funcs to be less
confusing

## Breaking Changes
- Remove the balance check from startup and the associated configs.
issue linked below

## Testing
- Updated some of the `Submitter`-specific tests to reflect changes to
the startup logic.

closes #1229