chore(sequencer)!: add metrics (#1248)
## Summary
This adds further metrics to the sequencer.

## Background
This should help diagnose block production slowdown when the sequencer
is stress-tested.

## Changes
- Added metrics (see below for list).
- Enabled `cnidarium` metrics.

Note that all histograms are still rendered as Prometheus summaries for
now. I have [an open PR](https://github.com/astriaorg/astria/pull/1192)
which will make it simple to provide buckets for histograms, after which
they will be rendered as true histograms.

## Testing
Testing will likewise be relatively simple once #1192 is merged. In the
meantime, I ran the smoke test using the code in this PR and manually
checked the new metrics are available and appear sane.

## Metrics
- Added `astria_sequencer_check_tx_duration_seconds` histograms with the
following labels:
    - `length check and parse raw tx`
    - `stateless check`
    - `nonce check`
    - `chain id check`
    - `balance check`
    - `check for removal`
    - `insert to app mempool`
- Added `astria_sequencer_actions_per_transaction_in_mempool` histogram
- Added `astria_sequencer_transaction_in_mempool_size_bytes` histogram
- Added `astria_sequencer_transactions_in_mempool_total` gauge
- Enabled `cnidarium_get_raw_duration_seconds` histogram
- Enabled `cnidarium_nonverifiable_get_raw_duration_seconds` histogram

## Related Issues
Closes #1247.