feat(cli): add cmd to collect withdrawal events and submit as actions (#1261)
## Summary
Adds `bridge collect-withdrawals` and `bridge submit-withdrawals`
subcommands to `astria-cli`.

## Background
The worker service `astria-bridge-withdrawer` is a closed system that
collects withdrawal event from the rollup and submits them to the
sequencer in a closed loop. But it can be desirable to inspect the
withdrawal events generated by the bridge contracts, and then submit
them manually. This functionality is now provided by `astria-cli`.

## Changes
- Add subcommands `bridge collect-withdrawal-s` to `astria-cli`, which
has two different modes of operation, depending on whether
`--to-rollup-height` is set:
- if set, it fetches all blocks between `--from-rollup-height` and
`--to-rollup-height` (inclusive), converts them to seuquencer actions,
and then writes them to a file.
- if not set, it fetches blocks from `--from-rollup-height` until SIGINT
(Ctrl-C) is received. After the signal, the converted actions are
written to a file.

## Testing
- Added `run-smoke-cli` in `charts/deploy.just` (@joroshiba)
- Added a `smoke-cli:` in `.github/workflows/docker-build.yaml`
(@joroshiba)

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>