chore(test): use a temporary file to not pollute the workspace (#1269)
## Summary
Uses a temporary file instead of writing to the workspace during smoke
tests.

## Background
It's bad practice to pollute the workspace when running tests. This
change writes the output to a temporary file instead, and ensures that
it is cleaned up after exit.

## Changes
- Update the `run-smoke-cli` recipe in `charts/deploy.just` to write to
a temporary generated by `mktemp` (without arguments to not run into
differences between Darwin/BSD and GNU/Linux)
- Use a trap to clean up the temporary on `EXIT`.

## Testing
Run `just run-smoke-cli`, observe that the workspace is clean.